refactor!: moves type definitions to where they're needed. stop exposing API BREAKING (#128)
## Description, Motivation and Context

- moves all type definitions into the `src/` tree as close to where they
define interface for (either within the modules themselves, or right
next to them)
- removes the `types`, `main` and `export` fields from package.json -
this is a command line utility, not a library. If there's a need for
this to be a library we can always bring it back. This is b.t.w.
theoretically 🚨 BREAKING, although I would be _highly_ surprised if
anyone was actually using VCO as a library.
- 🏕️ adds unit tests for `utensils` - they were already covered in
integration tests, but making it UT explicit will make them (it - it's
one function ...) easier during maintenance
- 🏕️ does a little shuffle with cli.ts, main.ts:
- cli.ts didn't contain the command line interface, it just ran it =>
renamed `cli.ts` -> `run-cli.ts` and adapted the package.json `bin`
field to reflect that
- main.ts contained both the command line interface definition _and_ the
main function -> split that into a `cli.ts` and a `main.ts`

## How Has This Been Tested?

- [x] green ci


## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation only change
- [x] Refactor (non-breaking change which fixes an issue without
changing functionality)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to change)